tests/pytest/requirements.txt
pytest>=8.0
pyyaml>=6.0
mitmproxy>=10.0
